Add tests for calc-size(calc-size(any, ...), size), which is allowed.
This change removes a TODO comment saying that the 'size' keyword should
be disallowed when the *nested* basis is 'any', and adds tests testing
that things work correctly for such cases.

This restriction is not required by the spec, and is not needed, since
the behavior of the 'size' keyword in this case is well-defined.

Bug: 313072
Change-Id: Icc2a879e8559ef5b0c97b852d179d73a7f113aa1